pyproject.toml
[tool.poetry]
name = "temperature-scoring"
version = "0.1.0"
description = "Temperature scoring according to SBTi tool"
authors = ["Patryk Kubiczek <patryk.kubiczek@instrat.pl>"]
readme = "README.md"
packages = [{include = "temperature_scoring"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.3.4"
openpyxl = "^3.0.9"
sbti-finance-tool = {git = "https://github.com/ScienceBasedTargets/SBTi-finance-tool"}


[tool.poetry.group.dev.dependencies]
requests = "^2.28.1"
icecream = "^2.1.3"
black = {version = "^22.12.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
